Rename WorldDataEmitter class to WorldView and update file structure
- Renamed renderer/viewer/lib/worldDataEmitter.ts to WorldView.ts
- Renamed WorldDataEmitter class to WorldView (kept WorldDataEmitterWorker unchanged)
- Updated all imports and references throughout the codebase:
  - src/watchOptions.ts
  - src/globals.d.ts
  - src/appViewer.ts
  - renderer/playground/baseScene.ts
  - renderer/viewer/three/panorama.ts
  - renderer/viewer/lib/worldrendererCommon.ts
- Maintained backward compatibility for WorldDataEmitterWorker class
