+++ src/globals.d.ts
@@ -12,7 +12,7 @@ declare const bot: Omit<import('mineflayer').Bot, 'world' | '_client'> & {
 }
 declare const __type_bot: typeof bot
 declare const appViewer: import('./appViewer').AppViewer
-declare const worldView: import('renderer/viewer/lib/worldDataEmitter').WorldDataEmitter | undefined
+declare const worldView: import('renderer/viewer/lib/WorldView').WorldView | undefined
 declare const addStatPerSec: (name: string) => void
 declare const localServer: import('flying-squid/dist/index').FullServer & { options } | undefined
 /** all currently loaded mc data */
